inline the header
